src/main/java/br/com/wtcode/qtorecebo/model/DescontoDependente.java
package br.com.wtcode.qtorecebo.model;

import java.math.BigDecimal;

import br.com.wtcode.qtorecebo.util.ManipulaProperties;

public class DescontoDependente{
	
	private ManipulaProperties manipulaProperties = new ManipulaProperties();
	
	public Salario calculaDesconto(Salario salario){
		if(salario.getNumeroDeDependentes() >0 && salario.getLiquido().compareTo(new BigDecimal("2000.00"))>=0  ){
			salario.setLiquido(salario.getLiquido().subtract(getValorDesconto()));
		}
		return salario;
	}

	private BigDecimal getValorDesconto() {
		return manipulaProperties.buscaAliquota("dependente.desconto");
	}

}
